% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markergrid-package.R
\docType{package}
\name{markergrid-package}
\alias{markergrid}
\alias{markergrid-package}
\title{markergrid: markergrid: Project features marking each area of a projection}
\description{
Create a grid of points with the specified grid_size. For each point on the grid, find its neighborhood by looking at the points in the dataset that are closest to it. For each feature in feature_names, calculate the specificity of that feature in the neighborhood of the grid point. This could be done using a metric such as the average distance between points that have that feature and points that do not have that feature. Display the feature name with the highest specificity for each grid point.
}
\author{
\strong{Maintainer}: Dimitri Meistermann \email{dimitri.meistermann@helsinki.fi} (\href{https://orcid.org/0000-0001-6614-2325}{ORCID})

}
\keyword{internal}
