% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{markergrid}
\alias{markergrid}
\title{Project features linked to a projection from each point on a grid}
\usage{
markergrid(
  proj,
  featureMat,
  gridResolution = 15,
  method = c("aurocFC", "auroc", "FC"),
  grid = NULL,
  transpose = TRUE,
  geomTextFun = shadowtext::geom_shadowtext,
  fontface = "bold.italic",
  returnPseudoSamples = FALSE,
  returnGrid = FALSE,
  ...
)
}
\arguments{
\item{proj}{Matrix or dataframe containing at least two vectors of numeric corresponding to the x/y coordinates.}

\item{featureMat}{A matrix with row as features (if \code{transpose}).}

\item{gridResolution}{Increase or decrease the number of points in the grid. It corresponds to the number of grid points along one axis if the projection is a square.}

\item{method}{Method for computing the feature specificity score. Can be \code{"aurocFC"}  (auroc * fold-change), \code{"auroc"} or \code{"FC"} (Fold-change).}

\item{grid}{A dataframe returned by this function if \code{returnGrid}. Contains data for plotting the grid (see \code{returnGrid} for more explanations)}

\item{transpose}{If \code{transpose}, samples are columns and features are rows.}

\item{geomTextFun}{Function used to plot the text if returnGrid=FALSE.}

\item{fontface}{Font face of the displayed features. Passed to the function specified in \code{geomTextFun}}

\item{returnPseudoSamples}{Return the convoluted samples that correspond to each point of the grid.}

\item{returnGrid}{Return a dataframe instead of plotting, corresponding to the coordinate of the grid,
Contains the following columns: \code{X} and \code{Y} (coordinates), \code{nContrib} (number of contributing observation to the grid point, the grid point is not created if \code{nContrib==0}),
\code{best} (best feature of the grid point), \code{score} (specificity score of the best feature), \code{sizeScore} (score weighted by number of observations, will determine the fontsize).}

\item{...}{additionnal parameters passed to the text plotting function-}
}
\value{
A ggproto object to add to a 2d projection by default. Can also be a dataframe if \code{returnGrid} or \code{returnPseudoSamples}.
}
\description{
Project features linked to a projection from each point on a grid
}
\details{
\enumerate{
\item Project the dataset on a 2D plane using a suitable algorithm (e.g. PCA or t-SNE).
\item Create a grid of points with the specified grid_size.
\item For each point on the grid, find its neighborhood by looking at the points in the dataset that are closest to it.
\item For each feature of the featureMat, calculate the specificity of that feature in the neighborhood of the grid point. This could be done using a metric such as the average distance between points that have that feature and points that do not have that feature.
\item Display the feature name with the highest specificity for each grid point.
}
}
\examples{
library(ggplot2)
library(oob)
data(iris)

qDataIris<-t(iris[,1:4])
irisUMAP<-UMAP(qDataIris)
ggproj<-ggplot(data.frame(x=irisUMAP[,1],y=irisUMAP[,2]),aes(x=x,y=y))+geom_point()+
  theme(panel.background = element_rect(fill = NA,
    colour = "black"), panel.grid.major = element_line(colour = NA),
    panel.grid.minor = element_line(colour = NA)) + guides(size = "none")
ggproj+markergrid(irisUMAP, qDataIris)+scale_size_continuous(range = c(1,3))
ggproj+markergrid(irisUMAP, qDataIris, geomTextFun=geom_text)+scale_size_continuous(range = c(1,3))
grid<-markergrid(irisUMAP,qDataIris, returnGrid=TRUE)
}
